Move fetch headers into its own file.
