+++ js/globals.ts
@@ -2,6 +2,7 @@
 import * as blob from "./blob";
 import * as console_ from "./console";
 import * as fetch_ from "./fetch";
+import { Headers } from "./headers";
 import { globalEval } from "./global_eval";
 import { libdeno } from "./libdeno";
 import * as textEncoding from "./text_encoding";
@@ -41,5 +42,5 @@ window.fetch = fetch_.fetch;
 
 // using the `as` keyword to mask the internal types when generating the
 // runtime library
-window.Headers = fetch_.Headers as domTypes.HeadersConstructor;
+window.Headers = Headers as domTypes.HeadersConstructor;
 window.Blob = blob.DenoBlob;
